add handle disburse reward function to utils
